controller: move shared informers in separate package
