+++ pkg/controller/shared/authorization_informers.go
@@ -1,4 +1,4 @@
-package controller
+package shared
 
 import (
 	"reflect"
