+++ pkg/controller/shared/doc.go
@@ -1,0 +1,2 @@
+// Package shared provides reusable support for shared informer implementations.
+package shared
