+++ pkg/controller/shared/legacy_kube_informers.go
@@ -1,4 +1,4 @@
-package controller
+package shared
 
 import (
 	"reflect"
