+++ pkg/controller/shared/shared_informer.go
@@ -1,4 +1,4 @@
-package controller
+package shared
 
 import (
 	"reflect"
